creating games view model and moving logic to it from the board games store
