+++ board_games_companion/lib/pages/games/games_page.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:board_games_companion/common/app_text.dart';
+import 'package:board_games_companion/pages/games/games_view_model.dart';
 import 'package:board_games_companion/widgets/common/slivers/bgc_sliver_header_delegate.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
@@ -19,7 +20,6 @@ import '../../models/navigation/playthroughs_page_arguments.dart';
 import '../../services/analytics_service.dart';
 import '../../services/rate_and_review_service.dart';
 import '../../stores/board_games_filters_store.dart';
-import '../../stores/board_games_store.dart';
 import '../../stores/user_store.dart';
 import '../../widgets/board_games/board_game_tile.dart';
 import '../../widgets/common/bgg_community_member_text_widget.dart';
@@ -34,14 +34,14 @@ import 'games_filter_panel.dart';
 
 class GamesPage extends StatefulWidget {
   const GamesPage(
-    this.boardGamesStore,
+    this.viewModel,
     this.userStore,
     this.analyticsService,
     this.rateAndReviewService, {
     Key? key,
   }) : super(key: key);
 
-  final BoardGamesStore boardGamesStore;
+  final GamesViewModel viewModel;
   final UserStore userStore;
   final AnalyticsService analyticsService;
   final RateAndReviewService rateAndReviewService;
@@ -58,26 +58,26 @@ class _GamesPageState extends State<GamesPage> with SingleTickerProviderStateMix
     _topTabController = TabController(
       length: 3,
       vsync: this,
-      initialIndex: widget.boardGamesStore.selectedTab.index,
+      initialIndex: widget.viewModel.selectedTab.index,
     );
     super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
-    if (widget.boardGamesStore.loadDataState == LoadDataState.loaded) {
-      if (!widget.boardGamesStore.anyBoardGamesInCollections &&
+    if (widget.viewModel.loadDataState == LoadDataState.loaded) {
+      if (!widget.viewModel.anyBoardGamesInCollections &&
           (widget.userStore.user?.name.isEmpty ?? true)) {
         return const _Empty();
       }
 
       return _Collection(
-        boardGamesStore: widget.boardGamesStore,
+        viewModel: widget.viewModel,
         topTabController: _topTabController,
         analyticsService: widget.analyticsService,
         rateAndReviewService: widget.rateAndReviewService,
       );
-    } else if (widget.boardGamesStore.loadDataState == LoadDataState.error) {
+    } else if (widget.viewModel.loadDataState == LoadDataState.error) {
       return const Center(
         child: GenericErrorMessage(),
       );
@@ -95,14 +95,14 @@ class _GamesPageState extends State<GamesPage> with SingleTickerProviderStateMix
 
 class _Collection extends StatelessWidget {
   const _Collection({
-    required this.boardGamesStore,
+    required this.viewModel,
     required this.topTabController,
     required this.analyticsService,
     required this.rateAndReviewService,
     Key? key,
   }) : super(key: key);
 
-  final BoardGamesStore boardGamesStore;
+  final GamesViewModel viewModel;
   final TabController topTabController;
   final AnalyticsService analyticsService;
   final RateAndReviewService rateAndReviewService;
@@ -114,47 +114,47 @@ class _Collection extends StatelessWidget {
       child: CustomScrollView(
         slivers: <Widget>[
           _AppBar(
-            boardGamesStore: boardGamesStore,
+            viewModel: viewModel,
             topTabController: topTabController,
             analyticsService: analyticsService,
             rateAndReviewService: rateAndReviewService,
             updateSearchResults: (String searchPhrase) => _updateSearchResults(searchPhrase),
           ),
-          if (boardGamesStore.collectionSate == CollectionState.emptySearchResult)
+          if (viewModel.collectionSate == CollectionState.emptySearchResult)
             _EmptySearchResult(
-              boardGamesStore: boardGamesStore,
+              viewModel: viewModel,
               onClearSearch: () => _updateSearchResults(''),
             ),
-          if (boardGamesStore.collectionSate == CollectionState.emptyCollection)
-            _EmptyCollection(boardGamesStore: boardGamesStore),
-          if (boardGamesStore.collectionSate == CollectionState.collection) ...[
+          if (viewModel.collectionSate == CollectionState.emptyCollection)
+            _EmptyCollection(gamesViewModel: viewModel),
+          if (viewModel.collectionSate == CollectionState.collection) ...[
             SliverPersistentHeader(
               pinned: true,
               delegate: BgcSliverHeaderDelegate(
                 title: sprintf(
                   AppText.gamesPageMainGamesSliverSectionTitleFormat,
-                  [boardGamesStore.totalMainGamesInCollections],
+                  [viewModel.totalMainGamesInCollections],
                 ),
               ),
             ),
             _Grid(
-              boardGames: boardGamesStore.mainGamesInCollections,
-              collectionType: boardGamesStore.selectedTab.toCollectionType(),
+              boardGames: viewModel.mainGamesInCollections,
+              collectionType: viewModel.selectedTab.toCollectionType(),
               analyticsService: analyticsService,
             ),
-            if (boardGamesStore.hasAnyExpansionsInSelectedCollection) ...[
+            if (viewModel.hasAnyExpansionsInSelectedCollection) ...[
               SliverPersistentHeader(
                 pinned: true,
                 delegate: BgcSliverHeaderDelegate(
                   title: sprintf(
                     AppText.gamesPageExpansionsSliverSectionTitleFormat,
-                    [boardGamesStore.totalExpansionsInCollections],
+                    [viewModel.totalExpansionsInCollections],
                   ),
                 ),
               ),
               _Grid(
-                boardGames: boardGamesStore.expansionsInSelectedCollection,
-                collectionType: boardGamesStore.selectedTab.toCollectionType(),
+                boardGames: viewModel.expansionsInSelectedCollection,
+                collectionType: viewModel.selectedTab.toCollectionType(),
                 analyticsService: analyticsService,
               ),
             ]
@@ -166,7 +166,7 @@ class _Collection extends StatelessWidget {
   }
 
   Future<void> _updateSearchResults(String searchPhrase) async {
-    boardGamesStore.updateSearchResults(searchPhrase);
+    viewModel.updateSearchResults(searchPhrase);
 
     await rateAndReviewService.increaseNumberOfSignificantActions();
   }
@@ -174,7 +174,7 @@ class _Collection extends StatelessWidget {
 
 class _AppBar extends StatefulWidget {
   const _AppBar({
-    required this.boardGamesStore,
+    required this.viewModel,
     required this.topTabController,
     required this.analyticsService,
     required this.rateAndReviewService,
@@ -182,7 +182,7 @@ class _AppBar extends StatefulWidget {
     Key? key,
   }) : super(key: key);
 
-  final BoardGamesStore boardGamesStore;
+  final GamesViewModel viewModel;
   final TabController topTabController;
   final AnalyticsService analyticsService;
   final RateAndReviewService rateAndReviewService;
@@ -210,7 +210,7 @@ class _AppBarState extends State<_AppBar> {
 
   @override
   Widget build(BuildContext context) {
-    if (widget.boardGamesStore.searchPhrase?.isEmpty ?? true) {
+    if (widget.viewModel.searchPhrase?.isEmpty ?? true) {
       _searchController.text = '';
     }
 
@@ -229,7 +229,7 @@ class _AppBarState extends State<_AppBar> {
         style: AppTheme.defaultTextFieldStyle,
         decoration: InputDecoration(
           hintText: 'Search for a game...',
-          suffixIcon: (widget.boardGamesStore.searchPhrase?.isNotEmpty ?? false)
+          suffixIcon: (widget.viewModel.searchPhrase?.isNotEmpty ?? false)
               ? IconButton(
                   icon: const Icon(
                     Icons.clear,
@@ -247,7 +247,7 @@ class _AppBarState extends State<_AppBar> {
         ),
         onSubmitted: (searchPhrase) async {
           _debounce?.cancel();
-          if (widget.boardGamesStore.searchPhrase != _searchController.text) {
+          if (widget.viewModel.searchPhrase != _searchController.text) {
             await widget.updateSearchResults(_searchController.text);
           }
           _searchFocusNode.unfocus();
@@ -260,7 +260,7 @@ class _AppBarState extends State<_AppBar> {
               icon: boardGamesFiltersStore.anyFiltersApplied
                   ? const Icon(Icons.filter_alt_rounded, color: AppTheme.accentColor)
                   : const Icon(Icons.filter_alt_outlined, color: AppTheme.accentColor),
-              onPressed: widget.boardGamesStore.allboardGames.isNotEmpty
+              onPressed: widget.viewModel.anyBoardGames
                   ? () async {
                       await _openFiltersPanel(context);
                       await widget.analyticsService.logEvent(name: Analytics.FilterCollection);
@@ -275,24 +275,24 @@ class _AppBarState extends State<_AppBar> {
         preferredSize: const Size.fromHeight(74),
         child: TabBar(
           onTap: (int index) {
-            widget.boardGamesStore.selectedTab = index.toGamesTab();
+            widget.viewModel.selectedTab = index.toGamesTab();
           },
           controller: widget.topTabController,
           tabs: <Widget>[
             _TopTab(
               'Owned',
               Icons.grid_on,
-              isSelected: widget.boardGamesStore.selectedTab == GamesTab.Owned,
+              isSelected: widget.viewModel.selectedTab == GamesTab.Owned,
             ),
             _TopTab(
               'Friends',
               Icons.group,
-              isSelected: widget.boardGamesStore.selectedTab == GamesTab.Friends,
+              isSelected: widget.viewModel.selectedTab == GamesTab.Friends,
             ),
             _TopTab(
               'Wishlist',
               Icons.card_giftcard,
-              isSelected: widget.boardGamesStore.selectedTab == GamesTab.Wishlist,
+              isSelected: widget.viewModel.selectedTab == GamesTab.Wishlist,
             ),
           ],
           indicatorColor: AppTheme.accentColor,
@@ -490,12 +490,12 @@ class _ImportDataFromBggSectionState extends State<_ImportDataFromBggSection> {
 
 class _EmptySearchResult extends StatelessWidget {
   const _EmptySearchResult({
-    required this.boardGamesStore,
+    required this.viewModel,
     required this.onClearSearch,
     Key? key,
   }) : super(key: key);
 
-  final BoardGamesStore boardGamesStore;
+  final GamesViewModel viewModel;
   final VoidCallback onClearSearch;
 
   @override
@@ -516,7 +516,7 @@ class _EmptySearchResult extends StatelessWidget {
                         '''It looks like you don't have any board games in your collection that match the search phrase ''',
                   ),
                   TextSpan(
-                    text: boardGamesStore.searchPhrase,
+                    text: viewModel.searchPhrase,
                     style: const TextStyle(
                       fontWeight: FontWeight.bold,
                     ),
@@ -545,10 +545,10 @@ class _EmptySearchResult extends StatelessWidget {
 class _EmptyCollection extends StatelessWidget {
   const _EmptyCollection({
     Key? key,
-    required this.boardGamesStore,
+    required this.gamesViewModel,
   }) : super(key: key);
 
-  final BoardGamesStore boardGamesStore;
+  final GamesViewModel gamesViewModel;
 
   @override
   Widget build(BuildContext context) {
@@ -572,13 +572,13 @@ class _EmptyCollection extends StatelessWidget {
                             text: "It looks like you don't have any board games in your ",
                           ),
                           TextSpan(
-                            text: boardGamesStore.selectedTab
+                            text: gamesViewModel.selectedTab
                                 .toCollectionType()
                                 .toHumandReadableText(),
                             style: const TextStyle(fontWeight: FontWeight.bold),
                           ),
                           const TextSpan(text: ' collection yet.'),
-                          if (boardGamesStore.selectedTab == GamesTab.Wishlist &&
+                          if (gamesViewModel.selectedTab == GamesTab.Wishlist &&
                               (userStore.user?.name.isNotEmpty ?? false)) ...[
                             const TextSpan(text: "\n\nIf you want to see board games from BGG's  "),
                             const TextSpan(
@@ -601,7 +601,7 @@ class _EmptyCollection extends StatelessWidget {
                       ),
                       textAlign: TextAlign.justify,
                     ),
-                    if (boardGamesStore.selectedTab == GamesTab.Wishlist &&
+                    if (gamesViewModel.selectedTab == GamesTab.Wishlist &&
                         (userStore.user?.name.isNotEmpty ?? false)) ...[
                       const SizedBox(height: Dimensions.doubleStandardSpacing),
                       Align(
