--- board_games_companion/lib/stores/board_game_details_in_collection_store.dart
@@ -1,33 +1,0 @@
-import 'package:collection/collection.dart' show IterableExtension;
-import 'package:flutter/cupertino.dart';
-
-import '../models/hive/board_game_details.dart';
-import 'board_games_store.dart';
-
-class BoardGameDetailsInCollectionStore extends ChangeNotifier {
-  BoardGameDetailsInCollectionStore(
-    this._boardGamesStore,
-    this._boardGameDetails,
-  );
-
-  final BoardGamesStore _boardGamesStore;
-  BoardGameDetails? _boardGameDetails;
-
-  bool get isInCollection {
-    if ((_boardGamesStore.filteredBoardGames?.isEmpty ?? true) || _boardGameDetails == null) {
-      return false;
-    }
-
-    final boardGameInCollection = _boardGamesStore.filteredBoardGames!.firstWhereOrNull(
-      (boardGameDetails) => boardGameDetails.id == _boardGameDetails!.id,
-    );
-    return boardGameInCollection != null;
-  }
-
-  void updateIsInCollectionStatus([BoardGameDetails? boardGameDetails]) {
-    if (boardGameDetails != null) {
-      _boardGameDetails = boardGameDetails;
-    }
-    notifyListeners();
-  }
-}
